+++ src/generateTypeFiles.js
